+++ Demo/Keyboard/Demo/DemoAutocompleteSuggestionProvider.swift
@@ -6,15 +6,42 @@
 //  Copyright © 2021 Daniel Saidi. All rights reserved.
 //
 
+import Foundation
 import KeyboardKit
 
 /**
  This demo provider simply returns the current word suffixed
  with "ly", "er" and "ter".
  
- This class is shared between the demo app and all keyboards.
+ The provider is registered by the input view controller, to
+ show you how to can register your own provoder. However, it
+ is overwritten by a `StandardAutocompleteSuggestionProvider`
+ from KeyboardKit Pro when the demo registers a pro license.
  */
 class DemoAutocompleteSuggestionProvider: AutocompleteSuggestionProvider {
+    
+    var locale: Locale = .current
+    
+    
+    var canIgnoreWords: Bool { false }
+    
+    var ignoredWords: [String] = []
+    
+    func hasIgnoredWord(_ word: String) -> Bool { false }
+    
+    func ignoreWord(_ word: String) {}
+    
+    func removeIgnoredWord(_ word: String) {}
+    
+    
+    var canLearnWords: Bool { false }
+    
+    func hasLearnedWord(_ word: String) -> Bool { false }
+    
+    func learnWord(_ word: String) {}
+    
+    func unlearnWord(_ word: String) {}
+    
     
     func autocompleteSuggestions(for text: String, completion: AutocompleteResponse) {
         guard text.count > 0 else { return completion(.success([])) }